server/requirements.txt
affine==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
color-operations==0.2.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.123.3
fonttools==4.61.0
geojson-pydantic==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
morecantile==6.2.0
numexpr==2.14.1
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pydantic==2.12.5
pydantic-settings>=2.1.0
pydantic_core==2.41.5
pyparsing==3.2.5
pyproj==3.7.2
pystac==1.14.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
rasterio==1.4.3
requests>=2.31.0
rio-tiler==7.9.2
simplejson==3.20.2
six==1.17.0
starlette==0.50.0
titiler-core==0.26.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
